feat/extract-minipool (#31)
* feat/extract-minipool
